feat: add feature flag support to go builds
